Move functions related to the clouds to a separate file
